requirements.txt
python-dotenv
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu